requirements.txt
# Core dependencies - versiuni actualizate și securizate
flask==3.0.0
python-telegram-bot==21.5
yt-dlp==2024.12.13
requests==2.31.0
werkzeug==3.0.1
Pillow==10.1.0
PyYAML==6.0.1

# Security dependencies
cryptography==41.0.8
python-dotenv==1.0.0

# Performance and utilities
aiofiles==23.2.1
psutil==5.9.6

# Testing dependencies (doar pentru development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-html==4.1.1
pytest-mock==3.12.0

# Rate limiting și monitoring
aiohttp==3.9.1

# Cleanup și optimizare
chardet==5.2.0
urllib3==2.1.0
